new(event_derive)!: derive(AnyEvent) macro
This lets us easily derive all the required trait impls for custom
enums (e.g. a subset of all the events for a particular plugin).
Also, it makes the implementation more symmetric (AnyEvent no longer
needs special support as a RawEvent payload).

+++ falco_event/src/events/payload.rs
@@ -40,7 +40,7 @@ pub enum PayloadFromBytesError {
 
     /// Unsupported event type
     #[error("unsupported event type {0}")]
-    UnsupportedEventType(u32),
+    UnsupportedEventType(u16),
 }
 
 pub trait PayloadToBytes {
